cast as int

+++ models/tmp/stg_stripe__invoice_line_items.sql
@@ -2,7 +2,7 @@ with invoice_line_items_period as (
     select
         _airbyte_invoice_line_items_hashid,
         to_timestamp_ntz("start", 0) as period_start,
-        to_timestamp_ntz(END, 0) as period_end
+        to_timestamp_ntz(END::int, 0) as period_end
     from {{ var('invoice_line_items_period') }}
 ),
